Fixed all compiler warnings

+++ wifi802_11.cpp
@@ -111,7 +111,7 @@ void wifi_802_receive_cb(void(*cb)(const uint8_t *, int, uint8_t)) {
 void wifi_802_11_send(const uint8_t *d, int len) {
   uint8_t buf[500];
   for(int i=0;i<5;i++){
-  if(len>sizeof(buf)-sizeof(raw_HEADER)-2) return;
+  if((unsigned long)len>sizeof(buf)-sizeof(raw_HEADER)-2) return;
 
   memcpy(buf,raw_HEADER, sizeof(raw_HEADER));
   memcpy(buf+sizeof(raw_HEADER)+2, d, len);
